requirements.txt
gradio
transformers
torch
pillow
numpy
huggingface_hub
peft
unsloth
bitsandbytes
